pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memorygraph"
version = "1.0.0"
description = "API for processing and storing content summaries"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi",
    "uvicorn",
    "arango",
    "pydantic",
    "python-dotenv"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"] 